LPS-96911 Move logic to base class so we can use elsewhere later
